requirements/dev.txt
-r base.txt
coverage
django-coverage
django-debug-toolbar
fabric
futures
ipdb
ipython
pysqlite
sphinx
tornado
